requirements.txt
selenium==4.15.0
groq==0.5.0
python-dotenv==1.0.0
requests==2.31.0
webdriver-manager==4.0.1
httpx==0.24.0